restructured repo into sub repos for kalman and e.t.#
