+++ firedrake_da/__init__.py
@@ -2,10 +2,13 @@ from __future__ import absolute_import
 
 from firedrake_da.utils import *  # noqa
 from firedrake_da.observations import *  # noqa
-from firedrake_da.cov import *  # noqa
-from firedrake_da.kalman import *  # noqa
 from firedrake_da.emd import *  # noqa
 from firedrake_da.localisation import *  # noqa
-from firedrake_da.weight_update import *  # noqa
-from firedrake_da.ensemble_transform import *  # noqa
+
+from firedrake_da.kalman.cov import *  # noqa
+from firedrake_da.kalman.kalman import *  # noqa
+
+from firedrake_da.ensemble_transform.weight_update import *  # noqa
+from firedrake_da.ensemble_transform.ensemble_transform import *  # noqa
+
 from firedrake_da.ml.coupling import *  # noqa
