+++ firedrake_da/ensemble_transform/__init__.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import
+
+from firedrake_da.ensemble_transform.ensemble_transform import *  # noqa
+from firedrake_da.ensemble_transform.weight_update import *  # noqa
