+++ firedrake_da/ensemble_transform/weight_update.py
@@ -10,7 +10,6 @@ from firedrake.mg.utils import get_level
 
 import numpy as np
 
-from firedrake_da.cov import *
 from firedrake_da.observations import *
 from firedrake_da.localisation import *
