+++ firedrake_da/kalman/__init__.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import
+
+from firedrake_da.kalman.kalman import *  # noqa
+from firedrake_da.kalman.cov import *  # noqa
