+++ tests/ensemble_transform/test_ensemble_transform.py
@@ -5,6 +5,7 @@ from __future__ import absolute_import
 from firedrake import *
 
 from firedrake_da import *
+from firedrake_da.ensemble_transform import *
 
 import numpy as np
